separated python API package into another directory
