+++ api/.idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/.idea/web-ui-nuxt.iml" filepath="$PROJECT_DIR$/.idea/web-ui-nuxt.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/pronunciation-api.iml" filepath="$PROJECT_DIR$/.idea/pronunciation-api.iml" />
     </modules>
   </component>
 </project>
