+++ api/pyproject.toml
@@ -1,0 +1,57 @@
+# Package Manager used: https://github.com/pdm-project/pdm
+[project]
+name = "pronunciation-api"
+version = "0.1.0"
+description = "Package for mispronunciation detection API (via FastAPI) based on machine learning model"
+
+authors = [
+    { name = "Maxim Ivanov",        email = "Moxeeeem@gmail.com" },
+    { name = "Mikhail Dorokhin",    email = "misha.dorohin5@gmail.com" },
+    { name = "Victor Khovanov",     email = "victorkhovtko@gmail.com" },
+]
+
+dependencies = []
+requires-python = ">=3.10"
+readme = "README.md"
+license = { text = "MIT" }
+
+
+# Additional configuration for poetry (https://python-poetry.org/docs/pyproject/) 
+[tool.poetry]
+name = "pronunciation-api"
+version = "0.1.0"
+description = "Package for mispronunciation detection API (via FastAPI) based on machine learning model"
+authors = [
+    "Maxim Ivanov <Moxeeeem@gmail.com>",
+    "Quasarel <misha.dorohin5@gmail.com>",
+    "Victor Khovanov <victorkhovtko@gmail.com>",
+]
+readme          = "README.md"
+license         = "MIT"
+package-mode    = true
+
+
+[tool.poetry.dependencies]
+python          = "^3.10"
+torch           = "^2.5.1"
+supabase        = "^2.10.0"
+fastapi         = { extras = ["standard"], version = "^0.115.6" }
+librosa         = "^0.10.2.post1"
+transformers    = "^4.46.2"
+phonemizer      = "^3.3.0"
+ortools         = "^9.11.4210"
+dtwalign        = "^0.1.1"
+
+
+[build-system]
+requires = ["pdm-backend"]
+build-backend = "pdm.backend"
+
+# (!) UNCOMMENT FOR USING POETRY for package building 
+# [build-system]
+# requires = ["poetry-core"]
+# build-backend = "poetry.core.masonry.api"
+
+
+[tool.pdm]
+distribution = true
