--- web-ui-nuxt/.idea/misc.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.11 (web-ui-nuxt) (2)" project-jdk-type="Python SDK" />
-</project>
