--- web-ui-nuxt/pyproject.toml
@@ -1,22 +1,0 @@
-[tool.poetry]
-name = "web-ui-nuxt"
-version = "0.1.0"
-description = ""
-authors = ["Quasarel <misha.dorohin5@gmail.com>"]
-readme = "README.md"
-
-[tool.poetry.dependencies]
-python = "^3.10"
-torch = "^2.5.1"
-supabase = "^2.10.0"
-fastapi = {extras = ["standard"], version = "^0.115.5"}
-librosa = "^0.10.2.post1"
-transformers = "^4.46.2"
-phonemizer = "^3.3.0"
-ortools = "^9.11.4210"
-dtwalign = "^0.1.1"
-
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
