--- web-ui-nuxt/requirements.txt
@@ -1,220 +1,0 @@
-absl-py==2.1.0
-aiohappyeyeballs==2.4.2
-aiohttp==3.10.10
-aiosignal==1.3.1
-aiostream==0.5.2
-annotated-types==0.7.0
-anyio==4.6.0
-argon2-cffi==23.1.0
-argon2-cffi-bindings==21.2.0
-arrow==1.3.0
-asgiref==3.8.1
-asttokens==2.4.1
-async-lru==2.0.4
-attrs==24.2.0
-audioread==3.0.1
-babel==2.16.0
-beautifulsoup4==4.12.3
-bibtexparser==2.0.0b7
-bleach==6.1.0
-certifi==2024.8.30
-cffi==1.17.1
-charset-normalizer==3.3.2
-click==8.1.7
-clldutils==3.22.2
-colorama==0.4.6
-colorlog==6.8.2
-comm==0.2.2
-contourpy==1.3.0
-csvw==3.3.1
-cycler==0.12.1
-datasets==3.0.1
-debugpy==1.8.6
-decorator==5.1.1
-defusedxml==0.7.1
-deprecation==2.1.0
-dill==0.3.8
-distro==1.9.0
-Django==5.0.6
-dlinfo==1.2.1
-dtwalign==0.1.1
-evaluate==0.4.3
-executing==2.1.0
-fastapi==0.110.2
-fastjsonschema==2.20.0
-feedparser==6.0.11
-filelock==3.16.1
-fonttools==4.54.1
-fqdn==1.5.1
-frozenlist==1.4.1
-fsspec==2024.6.1
-gotrue==2.9.3
-groq==0.11.0
-grpclib==0.4.7
-h11==0.14.0
-h2==4.1.0
-hpack==4.0.0
-httpcore==1.0.5
-httptools==0.6.4
-httpx==0.27.2
-huggingface-hub==0.25.1
-hyperframe==6.0.1
-idna==3.10
-immutabledict==4.2.0
-iniconfig==2.0.0
-ipykernel==6.29.5
-ipython==8.27.0
-ipywidgets==8.1.5
-isodate==0.6.1
-isoduration==20.11.0
-jedi==0.19.1
-Jinja2==3.1.4
-jiwer==3.0.4
-joblib==1.4.2
-json5==0.9.25
-jsonpointer==3.0.0
-jsonschema==4.23.0
-jsonschema-specifications==2023.12.1
-jupyter-events==0.10.0
-jupyter-lsp==2.2.5
-jupyter_client==8.6.3
-jupyter_core==5.7.2
-jupyter_server==2.14.2
-jupyter_server_terminals==0.5.3
-jupyterlab==4.2.5
-jupyterlab_pygments==0.3.0
-jupyterlab_server==2.27.3
-jupyterlab_widgets==3.0.13
-kiwisolver==1.4.7
-language-tags==1.2.0
-lazy_loader==0.4
-Levenshtein==0.26.0
-librosa==0.10.2.post1
-llvmlite==0.43.0
-lxml==5.3.0
-Markdown==3.7
-markdown-it-py==3.0.0
-MarkupSafe==2.1.5
-matplotlib==3.9.2
-matplotlib-inline==0.1.7
-mdurl==0.1.2
-mistune==3.0.2
-modal==0.64.151
-modal-client==0.64.151
-mpmath==1.3.0
-msgpack==1.1.0
-multidict==6.1.0
-multiprocess==0.70.16
-nbclient==0.10.0
-nbconvert==7.16.4
-nbformat==5.10.4
-nest-asyncio==1.6.0
-networkx==3.3
-notebook==7.2.2
-notebook_shim==0.2.4
-numba==0.60.0
-numpy==2.0.2
-ortools==9.11.4210
-overrides==7.7.0
-packaging==24.0
-pandas==2.2.3
-pandocfilters==1.5.1
-parso==0.8.4
-phonemizer==3.3.0
-pillow==10.4.0
-platformdirs==4.3.6
-pluggy==1.4.0
-pooch==1.8.2
-postgrest==0.17.2
-prometheus_client==0.21.0
-prompt_toolkit==3.0.48
-protobuf==5.26.1
-psutil==6.0.0
-pure_eval==0.2.3
-pyarrow==17.0.0
-PyAudio==0.2.14
-pycparser==2.22
-pydantic==2.9.2
-pydantic_core==2.23.4
-Pygments==2.18.0
-pylatexenc==2.10
-pyparsing==3.1.4
-pytest==8.1.1
-python-dateutil==2.9.0.post0
-python-dotenv==1.0.1
-python-json-logger==2.0.7
-python-Levenshtein==0.26.0
-python-multipart==0.0.17
-pytz==2024.2
-pywin32==306
-pywinpty==2.0.13
-PyYAML==6.0.2
-pyzmq==26.2.0
-RapidFuzz==3.10.0
-rdflib==7.0.0
-realtime==2.0.6
-referencing==0.35.1
-regex==2024.9.11
-requests==2.32.3
-rfc3339-validator==0.1.4
-rfc3986==1.5.0
-rfc3986-validator==0.1.1
-rich==13.8.1
-rpds-py==0.20.0
-sacremoses==0.1.1
-safetensors==0.4.5
-scikit-learn==1.5.2
-scipy==1.14.1
-seaborn==0.13.2
-segments==2.2.1
-Send2Trash==1.8.3
-sequence_align==0.2.0
-sgmllib3k==1.0.0
-shellingham==1.5.4
-sigtools==4.0.1
-six==1.16.0
-sniffio==1.3.1
-sounddevice==0.5.0
-soundfile==0.12.1
-soupsieve==2.5
-soxr==0.5.0.post1
-sqlparse==0.5.0
-stack-data==0.6.3
-starlette==0.37.2
-storage3==0.8.2
-supabase==2.9.1
-supafunc==0.6.2
-sympy==1.13.1
-synchronicity==0.7.7
-tabulate==0.9.0
-terminado==0.18.1
-threadpoolctl==3.5.0
-tinycss2==1.3.0
-tokenizers==0.20.1
-toml==0.10.2
-torch==2.5.1
-torchaudio==2.4.1
-torchvision==0.19.1
-tornado==6.4.1
-tqdm==4.66.5
-traitlets==5.14.3
-transformers==4.46.1
-typer==0.12.5
-types-certifi==2021.10.8.3
-types-python-dateutil==2.9.0.20240906
-types-toml==0.10.8.20240310
-typing_extensions==4.12.2
-tzdata==2024.1
-uri-template==1.3.0
-uritemplate==4.1.1
-urllib3==2.2.3
-uvicorn==0.29.0
-watchfiles==0.24.0
-wcwidth==0.2.13
-webcolors==24.8.0
-webencodings==0.5.1
-websocket-client==1.8.0
-websockets==13.1
-widgetsnbextension==4.0.13
-xxhash==3.5.0
-yarl==1.13.1
